Restructure to v4 project
