+++ PROJECT
@@ -1,6 +1,6 @@
 domain: tigera.io
 layout:
-- go.kubebuilder.io/v3
+- go.kubebuilder.io/v4
 projectName: operator
 repo: github.com/tigera/operator
 version: "3"
