+++ internal/controller/applicationlayer_controller.go
@@ -14,28 +14,29 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers
+package controller
 
 import (
 	"github.com/go-logr/logr"
+
+	"github.com/tigera/operator/pkg/controller/applicationlayer"
+	"github.com/tigera/operator/pkg/controller/options"
+
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	"github.com/tigera/operator/pkg/controller/monitor"
-	"github.com/tigera/operator/pkg/controller/options"
 )
 
-// MonitorReconciler reconciles a Monitor object
-type MonitorReconciler struct {
+// ApplicationLayerReconciler reconciles a ApplicationLayer object
+type ApplicationLayerReconciler struct {
 	client.Client
 	Log    logr.Logger
 	Scheme *runtime.Scheme
 }
 
-// +kubebuilder:rbac:groups=operator.tigera.io,resources=monitors,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups=operator.tigera.io,resources=monitors/status,verbs=get;update;patch
+// +kubebuilder:rbac:groups=operator.tigera.io,resources=applicationlayers,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=operator.tigera.io,resources=applicationlayers/status,verbs=get;update;patch
 
-func (r *MonitorReconciler) SetupWithManager(mgr ctrl.Manager, opts options.AddOptions) error {
-	return monitor.Add(mgr, opts)
+func (r *ApplicationLayerReconciler) SetupWithManager(mgr ctrl.Manager, opts options.AddOptions) error {
+	return applicationlayer.Add(mgr, opts)
 }
