+++ internal/controller/logcollector_controller.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers
+package controller
 
 import (
 	"github.com/go-logr/logr"
@@ -22,32 +22,32 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/tigera/operator/pkg/controller/clusterconnection"
+	"github.com/tigera/operator/pkg/controller/logcollector"
 	"github.com/tigera/operator/pkg/controller/options"
 )
 
-// ManagementClusterConnectionReconciler reconciles a ManagementClusterConnection object
-type ManagementClusterConnectionReconciler struct {
+// LogCollectorReconciler reconciles a LogCollector object
+type LogCollectorReconciler struct {
 	client.Client
 	Log    logr.Logger
 	Scheme *runtime.Scheme
 }
 
-// +kubebuilder:rbac:groups=operator.tigera.io,resources=managementclusterconnections,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups=operator.tigera.io,resources=managementclusterconnections/status,verbs=get;update;patch
+// +kubebuilder:rbac:groups=operator.tigera.io,resources=logcollectors,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=operator.tigera.io,resources=logcollectors/status,verbs=get;update;patch
 
-//func (r *ManagementClusterConnectionReconciler) Reconcile(req ctrl.Request) (ctrl.Result, error) {
+//func (r *LogCollectorReconciler) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 //	_ = context.Background()
-//	_ = r.Log.WithValues("managementclusterconnection", req.NamespacedName)
+//	_ = r.Log.WithValues("logcollector", req.NamespacedName)
 //
 //	// your logic here
 //
 //	return ctrl.Result{}, nil
 //}
 
-func (r *ManagementClusterConnectionReconciler) SetupWithManager(mgr ctrl.Manager, opts options.AddOptions) error {
-	return clusterconnection.Add(mgr, opts)
-	//return ctrl.NewControllerManagedBy(mgr).
-	//	For(&operatorv1.ManagementClusterConnection{}).
+func (r *LogCollectorReconciler) SetupWithManager(mgr ctrl.Manager, opts options.AddOptions) error {
+	return logcollector.Add(mgr, opts)
+	// return ctrl.NewControllerManagedBy(mgr).
+	//	For(&operatorv1.LogCollector{}).
 	//	Complete(r)
 }
