+++ internal/controller/monitor_controller.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2022-2024 Tigera, Inc. All rights reserved.
+// Copyright (c) 2021-2024 Tigera, Inc. All rights reserved.
 /*
 
 Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,23 +14,28 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers
+package controller
 
 import (
 	"github.com/go-logr/logr"
-	"github.com/tigera/operator/pkg/controller/options"
-	"github.com/tigera/operator/pkg/controller/tiers"
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/tigera/operator/pkg/controller/monitor"
+	"github.com/tigera/operator/pkg/controller/options"
 )
 
-type TiersReconciler struct {
+// MonitorReconciler reconciles a Monitor object
+type MonitorReconciler struct {
 	client.Client
 	Log    logr.Logger
 	Scheme *runtime.Scheme
 }
 
-func (r *TiersReconciler) SetupWithManager(mgr ctrl.Manager, opts options.AddOptions) error {
-	return tiers.Add(mgr, opts)
+// +kubebuilder:rbac:groups=operator.tigera.io,resources=monitors,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=operator.tigera.io,resources=monitors/status,verbs=get;update;patch
+
+func (r *MonitorReconciler) SetupWithManager(mgr ctrl.Manager, opts options.AddOptions) error {
+	return monitor.Add(mgr, opts)
 }
