+++ internal/controller/packetcapture_controller.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2021-2024 Tigera, Inc. All rights reserved.
+// Copyright (c) 2024 Tigera, Inc. All rights reserved.
 /*
 
 Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,29 +14,28 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers
+package controller
 
 import (
 	"github.com/go-logr/logr"
-
-	"github.com/tigera/operator/pkg/controller/applicationlayer"
-	"github.com/tigera/operator/pkg/controller/options"
-
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/tigera/operator/pkg/controller/options"
+	"github.com/tigera/operator/pkg/controller/packetcapture"
 )
 
-// ApplicationLayerReconciler reconciles a ApplicationLayer object
-type ApplicationLayerReconciler struct {
+// PacketCaptureReconciler reconciles a PacketCapture object.
+type PacketCaptureReconciler struct {
 	client.Client
 	Log    logr.Logger
 	Scheme *runtime.Scheme
 }
 
-// +kubebuilder:rbac:groups=operator.tigera.io,resources=applicationlayers,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups=operator.tigera.io,resources=applicationlayers/status,verbs=get;update;patch
+// +kubebuilder:rbac:groups=operator.tigera.io,resources=packetcaptureapi,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=operator.tigera.io,resources=packetcaptureapi/status,verbs=get;update;patch
 
-func (r *ApplicationLayerReconciler) SetupWithManager(mgr ctrl.Manager, opts options.AddOptions) error {
-	return applicationlayer.Add(mgr, opts)
+func (pc *PacketCaptureReconciler) SetupWithManager(mgr ctrl.Manager, opts options.AddOptions) error {
+	return packetcapture.Add(mgr, opts)
 }
