+++ internal/controller/suite_test.go.generated
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers
+package controller
 
 // This file was generated by the v1.0.1 operator-sdk and it would
 // be nice to understand and start using this for testing. My understanding
